+++ src/imio/smartweb/core/contents/blocks/link/content.py
@@ -1,22 +1,15 @@
 # -*- coding: utf-8 -*-
 
+from imio.smartweb.core.contents.blocks.link.fields import IconsRadioWidget
+from imio.smartweb.core.contents.blocks.link.fields import LinkField
 from imio.smartweb.locales import SmartwebMessageFactory as _
+from plone.app.contenttypes.interfaces import ILink
+from plone.autoform import directives
 from plone.dexterity.content import Container
 from plone.namedfile.field import NamedBlobImage
-from zope.interface import implementer
+from plone.supermodel import model
 from zope import schema
-from zope.schema import TextLine
-from zope.schema.interfaces import ITextLine
-from plone.app.contenttypes.interfaces import ILink
-
-
-class ILinkField(ITextLine):
-    """ """
-
-
-@implementer(ILinkField)
-class LinkField(TextLine):
-    """ """
+from zope.interface import implementer
 
 
 class IBlockLink(ILink):
@@ -34,8 +27,10 @@ class IBlockLink(ILink):
 
     image = NamedBlobImage(title=_("Image"), required=False)
 
+    model.fieldset("icons", label=_("Icons"), fields=["svg_icon"])
+    directives.widget("svg_icon", IconsRadioWidget)
     svg_icon = schema.Choice(
-        title=_("Icon"),
+        title=_("Link illustration icon"),
         description=_("Only used in table view (takes precedence over image)"),
         source="imio.smartweb.vocabulary.Icons",
         required=False,